refactor: extract the hardened scratch-worktree harness into gitscratch
grist hardens every git invocation it makes: rebase.updateRefs=false so a
replay cannot rewrite the branches being simulated, rerere.enabled=false so a
simulated resolution never poisons the shared rr-cache, core.hooksPath aimed at
an empty directory, GIT_EDITOR=true, gc.auto=0, and a Drop that removes the
scratch worktree by path while deliberately never pruning. Those guards were
each validated by mutation and are pinned by tests.

The grime and grind shell functions do the same job with none of it, and every
guard added to grist from here on would be a guard they silently do not get.
Rewriting them as binaries only helps if there is exactly one implementation of
"safely dry-run git in a scratch worktree" for all three tools to share, so
Git, Scratch, the conflict newtypes, and the rebase replay move into a new
gitscratch crate. A scratch worktree can now only be built through
Scratch::create, and a Scratch only hands out a Git that already carries the
whole safety configuration — there is no way to get the worktree without also
getting the hardening.

Pure move, no behavior change. Scratch::replay_onto becomes
Scratch::replay_rebase (it never touched the Simulator), the private Cost
becomes the public Conflicts result type, and OrderingScore stays in grist
because ranking is grist's alone. grist re-exports the moved newtypes so its
public API is byte-for-byte what it was; its existing suite staying green is
the verification, per the exception stated in the design spec.

Refs #327
Slice 1 of specs/2026-07-30-grime-grind-design.md
