+++ Cargo.toml
@@ -131,6 +131,7 @@ buildinfo = { path = "src/buildinfo" }
 op-cache = { path = "src/op-cache" }
 clipboardmon = { path = "src/clipboardmon" }
 filewalker = { path = "src/filewalker" }
+gitscratch = { path = "src/gitscratch" }
 portplz-core = { path = "src/portplz-core" }
 repowalker = { path = "src/repowalker" }
 shellsetup = { path = "src/shellsetup" }
