+++ src/gitscratch/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "gitscratch"
+version = "0.1.0"
+edition.workspace = true
+description = "The hardened scratch-worktree harness for dry-running git operations without touching a real repository"
+
+[dependencies]
+anyhow.workspace = true
+tempfile.workspace = true
+
+[lints]
+workspace = true
