+++ src/grist/src/simulate.rs
@@ -12,29 +12,24 @@
 //! simulated main and then squashes the result in. Conflicts are counted, then
 //! resolved by staging the conflict markers verbatim.
 //!
-//! # Why markers, and what that means for the numbers
+//! # Where the replay itself lives
 //!
-//! Staging markers is the conservative auto-resolution: unlike `--ours` or
-//! `--theirs` it never silently discards a side. It does mean a later commit
-//! touching the same region conflicts again - but that is faithful to reality,
-//! since a human resolution also leaves later commits conflicting against the
-//! resolved state. Treat the totals as a cost index for comparing orderings
-//! measured under identical rules, not as an exact prediction.
+//! The scratch worktree, the pinned safety configuration, and the
+//! marker-staging resolution loop belong to [`gitscratch::Scratch`]. See
+//! [`gitscratch::scratch`] for why markers are staged and what that means for
+//! the numbers: the totals are a cost index for comparing orderings measured
+//! under identical rules, not an exact prediction. `grist` adds the squash and
+//! the ranking on top.
 
 use std::collections::{BTreeSet, HashMap};
 use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result};
-use tempfile::TempDir;
+use gitscratch::{BranchName, Conflicts, Git, Scratch};
 
-use crate::git::Git;
-use crate::metrics::{BranchName, Files, Hunks, OrderingScore, Stops};
+use crate::metrics::OrderingScore;
 use crate::plan::{ordering_count, permutations};
 use crate::rank::rank;
-
-/// Upper bound on rebase resolution rounds per branch, so a git state we failed
-/// to anticipate stalls the run instead of spinning forever.
-const MAX_RESOLUTION_ROUNDS: usize = 1_000;
 
 /// Orderings grow factorially. Seven branches is 5,040 replays - far past the
 /// point where waiting for an answer beats just picking one and rebasing.
@@ -133,7 +128,7 @@ impl Simulator {
     pub fn score(&self, order: &[BranchName]) -> Result<OrderingScore> {
         let scratch = Scratch::create(&self.repo, &self.base)?;
         let mut simulated_main = scratch.git().rev_parse(&self.base)?;
-        let mut total = Cost::default();
+        let mut total = Conflicts::default();
 
         for branch in order {
             let (next_main, step) = self.land(&scratch, &simulated_main, branch)?;
@@ -141,7 +136,7 @@ impl Simulator {
             total.absorb(step);
         }
 
-        Ok(total.into_score(order.to_vec()))
+        Ok(into_score(&total, order.to_vec()))
     }
 
     /// Score every order `branches` could land in, cheapest first.
@@ -161,8 +156,8 @@ impl Simulator {
         let scratch = Scratch::create(&self.repo, &self.base)?;
         let base_commit = scratch.git().rev_parse(&self.base)?;
 
-        let mut memo: HashMap<Vec<BranchName>, (String, Cost)> = HashMap::new();
-        memo.insert(Vec::new(), (base_commit, Cost::default()));
+        let mut memo: HashMap<Vec<BranchName>, (String, Conflicts)> = HashMap::new();
+        memo.insert(Vec::new(), (base_commit, Conflicts::default()));
 
         let mut scores = Vec::new();
 
@@ -190,7 +185,7 @@ impl Simulator {
             let (_, total) = memo
                 .get(&prefix)
                 .context("internal error: the full ordering was not simulated")?;
-            scores.push(total.clone().into_score(ordering));
+            scores.push(into_score(total, ordering));
         }
 
         Ok(rank(scores))
@@ -198,7 +193,12 @@ impl Simulator {
 
     /// Rebase `branch` onto `onto` and squash it in, reporting the new
     /// simulated main and what the step cost.
-    fn land(&self, scratch: &Scratch, onto: &str, branch: &BranchName) -> Result<(String, Cost)> {
+    fn land(
+        &self,
+        scratch: &Scratch,
+        onto: &str,
+        branch: &BranchName,
+    ) -> Result<(String, Conflicts)> {
         let git = scratch.git();
 
         self.report(&format!("replaying {branch}"));
@@ -207,84 +207,23 @@ impl Simulator {
         git.run(&["checkout", "-q", "--detach", branch.as_str()])
             .with_context(|| format!("could not check out '{branch}'"))?;
 
-        let cost = self.replay_onto(&git, scratch.path(), onto, branch)?;
+        let cost = scratch
+            .replay_rebase(onto)
+            .with_context(|| format!("could not replay '{branch}' onto the simulated main"))?;
         let next_main = squash_into(&git, onto, branch)?;
 
         Ok((next_main, cost))
     }
-
-    /// Rebase the checked-out branch onto `onto`, resolving as it goes.
-    fn replay_onto(
-        &self,
-        git: &Git,
-        worktree: &Path,
-        onto: &str,
-        branch: &BranchName,
-    ) -> Result<Cost> {
-        let mut cost = Cost::default();
-        let mut outcome = git.try_run(&["rebase", onto])?;
-
-        for _ in 0..MAX_RESOLUTION_ROUNDS {
-            if !rebase_in_progress(git, worktree)? {
-                anyhow::ensure!(
-                    outcome.success,
-                    "rebasing '{branch}' onto the simulated main failed without leaving a \
-                     rebase to resolve:\n{}\n{}",
-                    outcome.stdout,
-                    outcome.stderr
-                );
-                return Ok(cost);
-            }
-
-            let conflicted = git.lines(&["diff", "--name-only", "--diff-filter=U"])?;
-
-            if conflicted.is_empty() {
-                // The rebase halted without unmerged paths - typically a commit
-                // that became empty once its changes were already present.
-                // Nothing for a human to resolve, so it costs nothing.
-                outcome = git.try_run(&["rebase", "--skip"])?;
-                continue;
-            }
-
-            cost.stops += 1;
-            for file in conflicted {
-                cost.hunks += count_conflict_hunks(&worktree.join(&file))?;
-                cost.files.insert(file);
-            }
-
-            git.run(&["add", "-A"])?;
-            outcome = git.try_run(&["rebase", "--continue"])?;
-        }
-
-        anyhow::bail!("gave up rebasing '{branch}' after {MAX_RESOLUTION_ROUNDS} resolution rounds")
-    }
 }
 
-/// What replaying one branch - or a whole prefix of them - cost.
-#[derive(Default, Clone)]
-struct Cost {
-    stops: usize,
-    hunks: usize,
-    files: BTreeSet<String>,
-}
-
-impl Cost {
-    /// Fold another step's cost into this running total.
-    fn absorb(&mut self, other: Self) {
-        self.stops += other.stops;
-        self.hunks += other.hunks;
-        self.files.extend(other.files);
-    }
-
-    /// Attribute this cost to the ordering that produced it.
-    fn into_score(self, order: Vec<BranchName>) -> OrderingScore {
-        OrderingScore::new(
-            order,
-            Stops::new(self.stops),
-            Files::new(self.files.len()),
-            Hunks::new(self.hunks),
-        )
-    }
+/// Attribute a replay's cost to the ordering that produced it.
+fn into_score(conflicts: &Conflicts, order: Vec<BranchName>) -> OrderingScore {
+    OrderingScore::new(
+        order,
+        conflicts.stops(),
+        conflicts.files(),
+        conflicts.hunks(),
+    )
 }
 
 /// Collapse the checked-out (already rebased) branch into a single commit on
@@ -299,123 +238,4 @@ fn squash_into(git: &Git, parent: &str, branch: &BranchName) -> Result<String> {
         "-m",
         &format!("squash {branch}"),
     ])
-}
-
-/// Whether git is sitting in a halted rebase.
-fn rebase_in_progress(git: &Git, worktree: &Path) -> Result<bool> {
-    for state_dir in ["rebase-merge", "rebase-apply"] {
-        let path = git.run(&["rev-parse", "--git-path", state_dir])?;
-        if worktree.join(path).exists() {
-            return Ok(true);
-        }
-    }
-    Ok(false)
-}
-
-/// Count the conflict regions a human would have to hand-merge in one file.
-///
-/// Conflicts with no markers at all - binary files, add/add on a blob git will
-/// not diff, delete/modify - still cost one decision each.
-fn count_conflict_hunks(path: &Path) -> Result<usize> {
-    let Ok(contents) = std::fs::read(path) else {
-        // A delete/modify conflict can leave no file on disk; it is still one
-        // decision for the person resolving it.
-        return Ok(1);
-    };
-
-    let markers = contents
-        .split(|byte| *byte == b'\n')
-        .filter(|line| line.starts_with(b"<<<<<<<"))
-        .count();
-
-    Ok(markers.max(1))
-}
-
-/// A detached scratch worktree that removes itself.
-struct Scratch {
-    repo: PathBuf,
-    /// Never read: held solely so the temporary directory - and everything the
-    /// simulation wrote into it - is removed when the `Scratch` is dropped.
-    #[expect(dead_code, reason = "held only so the TempDir is removed on drop")]
-    dir: TempDir,
-    worktree: PathBuf,
-    /// Validated once in [`Scratch::create`] so every `Git` built from it can
-    /// have the path infallibly. An empty `core.hooksPath` is not "hooks off" -
-    /// git still resolves hook lookups against it - so a path that cannot be
-    /// spelled for git has to fail the run, not degrade into one.
-    hooks: String,
-}
-
-impl Scratch {
-    /// Add a detached worktree at `base` in a private temporary directory.
-    fn create(repo: &Path, base: &str) -> Result<Self> {
-        let dir = TempDir::new().context("could not create a scratch directory")?;
-        let worktree = dir.path().join("worktree");
-        let hooks_dir = dir.path().join("hooks");
-        std::fs::create_dir(&hooks_dir).context("could not create the empty hooks directory")?;
-        let hooks = hooks_dir
-            .to_str()
-            .context("scratch hooks path is not valid UTF-8")?
-            .to_owned();
-
-        let scratch = Self {
-            repo: repo.to_path_buf(),
-            dir,
-            worktree,
-            hooks,
-        };
-
-        scratch.repo_git().run(&[
-            "worktree",
-            "add",
-            "-q",
-            "--detach",
-            scratch.worktree_arg()?,
-            base,
-        ])?;
-
-        Ok(scratch)
-    }
-
-    fn path(&self) -> &Path {
-        &self.worktree
-    }
-
-    fn worktree_arg(&self) -> Result<&str> {
-        self.worktree
-            .to_str()
-            .context("scratch worktree path is not valid UTF-8")
-    }
-
-    /// A runner rooted in the real repository.
-    fn repo_git(&self) -> Git {
-        Git::new(&self.repo, self.hooks.as_str())
-    }
-
-    /// A runner rooted in the scratch worktree.
-    fn git(&self) -> Git {
-        Git::new(&self.worktree, self.hooks.as_str())
-    }
-}
-
-impl Drop for Scratch {
-    fn drop(&mut self) {
-        // Best effort: the TempDir goes away regardless, but git also keeps
-        // administrative state in the real repo that must be cleaned up.
-        // Removing by path takes both, and it runs while the TempDir is still
-        // alive, so git still sees the worktree it is being asked about.
-        //
-        // Deliberately no `worktree prune` afterwards. Pruning is repo-wide and
-        // immediate: it deletes the administrative state - including any halted
-        // rebase - of every worktree whose directory is merely *missing right
-        // now*, which is the normal condition for a worktree on an unmounted
-        // drive or a sleeping network mount. A dry run must not cost the
-        // developer a worktree. If the removal above ever fails, the leftover
-        // entry is inert, and git's own gc clears it once it ages out.
-        if let Ok(path) = self.worktree_arg() {
-            let _ = self
-                .repo_git()
-                .try_run(&["worktree", "remove", "--force", path]);
-        }
-    }
 }
